# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NanTuPy API",
    "version": "1.0.0",
    "description": "社交平台全量接口文档 — 自动生成\n\n认证方式：除注册/登录外，所有接口需在请求头携带 `Authorization: Bearer <jwt_token>`"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "本地开发服务器"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "用户注册",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "description": "请求体"
        },
        "security": []
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "用户登录（支持用户名或邮箱）",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "description": "请求体"
        },
        "security": []
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "获取当前登录用户信息",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "summary": "获取当前用户信息（/profile 别名）",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "修改当前用户个人资料",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/UserResponse"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "description": "请求体"
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "summary": "修改当前用户密码",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "description": "请求体"
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/auth/users/{user_id}": {
      "get": {
        "summary": "获取用户公开信息（可选认证）",
        "tags": [
          "Auth",
          "Users"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "username": {
                      "type": "string"
                    },
                    "display_name": {
                      "type": "string",
                      "nullable": true
                    },
                    "bio": {
                      "type": "string",
                      "nullable": true
                    },
                    "avatar_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "cover_photo_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "用户 ID"
          }
        ],
        "security": []
      }
    },
    "/api/auth/account": {
      "delete": {
        "summary": "注销当前用户账号",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "summary": "发送密码重置验证码",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "description": "6位验证码（开发环境直接返回）"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "description": "请求体"
        },
        "security": []
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "summary": "验证令牌并重置密码",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "description": "请求体"
        },
        "security": []
      }
    },
    "/api/auth/privacy": {
      "get": {
        "summary": "获取当前用户隐私设置",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile_visibility": {
                      "type": "string"
                    },
                    "post_default_visibility": {
                      "type": "string"
                    },
                    "show_email": {
                      "type": "boolean"
                    },
                    "allow_search": {
                      "type": "boolean"
                    },
                    "allow_friend_requests": {
                      "type": "string"
                    },
                    "notify_push": {
                      "type": "boolean"
                    },
                    "notify_message": {
                      "type": "boolean"
                    },
                    "notify_sound": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "更新当前用户隐私设置",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivacySettingsUpdateRequest"
              }
            }
          },
          "description": "请求体"
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "刷新 Token",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名，3-30字符"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "邮箱"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "密码，至少8位"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱或用户名"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "密码"
          }
        },
        "required": [
          "password"
        ]
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "access_token": {
            "type": "string",
            "nullable": true
          },
          "user_id": {
            "type": "integer",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "type": "object",
            "nullable": true,
            "description": "用户信息对象"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "display_name": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "avatar_url": {
            "type": "string",
            "nullable": true
          },
          "cover_photo_url": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ProfileUpdateRequest": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "avatar_url": {
            "type": "string",
            "nullable": true
          },
          "cover_photo_url": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "old_password": {
            "type": "string",
            "format": "password"
          },
          "new_password": {
            "type": "string",
            "format": "password"
          }
        },
        "required": [
          "old_password",
          "new_password"
        ]
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "6位验证码"
          },
          "new_password": {
            "type": "string",
            "format": "password"
          }
        },
        "required": [
          "token",
          "new_password"
        ]
      },
      "PrivacySettingsUpdateRequest": {
        "type": "object",
        "properties": {
          "profile_visibility": {
            "type": "string",
            "nullable": true,
            "enum": [
              "public",
              "friends_only",
              "private"
            ]
          },
          "post_default_visibility": {
            "type": "string",
            "nullable": true,
            "enum": [
              "public",
              "friends_only",
              "private",
              "custom"
            ]
          },
          "show_email": {
            "type": "boolean",
            "nullable": true
          },
          "allow_search": {
            "type": "boolean",
            "nullable": true
          },
          "allow_friend_requests": {
            "type": "string",
            "nullable": true,
            "enum": [
              "everyone",
              "friends_of_friends",
              "none"
            ]
          }
        }
      }
    }
  }
}